requirements.txt
scrapy
google-cloud-logging
